*	Changed the index related variables to size_t *	Made the domain error work again

+++ CircularDelay.hpp
@@ -1,5 +1,6 @@
 
-#include <exception>
+#include <stdexcept>
+#include <cstdlib>
 /**
  * @brief A class that functions as a smaple buffer.
  * @details You can use this to insert samples and use the get fuction to get a sample of the past.
@@ -12,10 +13,10 @@ class CircularDelay{
 public:
 	CircularDelay();
 	void push(type input);
-	type get(unsigned int delay);
+	type get(size_t delay);
 private:
 	type data[size];
-	unsigned int index;
+	size_t index;
 };
 
 /**
@@ -23,7 +24,7 @@ private:
  */
 template<typename type, size_t size>
 CircularDelay<type, size>::CircularDelay():index(size-1){
-	for (unsigned int i = 0; i < size; ++i){
+	for (size_t i = 0; i < size; ++i){
 		data[i] = 0;
 	}
 }
@@ -47,7 +48,7 @@ void CircularDelay<type, size>::push(type input){
  * @return The sample of delay ago.
  */
 template<typename type, size_t size>
-type CircularDelay<type, size>::get(unsigned int delay){
+type CircularDelay<type, size>::get(size_t delay){
 	if(delay >= size)
 		throw(std::domain_error("Tried to get a value that is longer ago than the size of a CircularDelay."));
 	return data[(index + size - delay) % size];
